Move parse_args function to cli_utils module
